requirements.txt
Flask==2.1.1
gunicorn==20.1.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
uvicorn==0.22.0
waitress==2.1.2
Werkzeug==2.3.4
click==8.1.3
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
uvicorn==0.22.0
waitress==2.1.2
Werkzeug==2.3.4
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.12.2
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdownify==0.11.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
outcome==1.2.0
packaging==23.2
perplexity-api @ git+https://gitlab.com/robowaifudev/perplexity-api@0baebed39beeedc13e6281af5505f4c4221fdb5f
Pillow==10.1.0
PySocks==1.7.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
selenium==4.12.0
sortedcontainers==2.4.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
trio==0.22.2
trio-websocket==0.10.4
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0